Move loading functions to new module file load_utils.py
