+++ src/viqa/meson.build
@@ -3,6 +3,7 @@ python_sources = [
     'kernels.py',
     '_metrics.py',
     'deprecation.py',
+    'load_utils.py',
     'utils.py',
     'fusion.py',
     'batch_mode.py',
