+++ src/viqa/nr_metrics/snr.py
@@ -36,7 +36,8 @@ from warnings import warn
 import numpy as np
 
 from viqa._metrics import NoReferenceMetricsInterface
-from viqa.utils import _visualize_snr_2d, _visualize_snr_3d, load_data
+from viqa.load_utils import load_data
+from viqa.utils import _visualize_snr_2d, _visualize_snr_3d
 
 
 class SNR(NoReferenceMetricsInterface):
